refactor(analysis): extract stateless query helpers into AnalyzerHelpers
Split Analyzer into the public tool methods (the what) and the
index-only query primitives — rendering, symbol/scope/package lookups,
glob predicates, range math, hierarchy, local-binding rendering and
implicit resolution (the how).
